Se separa el proyecto en 2 directorios, uno para el front y el otro para el backend, se crea el front con React y se testaron todos los endpoint, se utilizo el carrito con ID 2 para visualizar en la web los productos almacenados

+++ server/src/routes/ProductosRoutes.js
@@ -6,6 +6,8 @@ const container = new Contenedor('./src/data/productos.json')
 
 const routerProductos = Router()
 
+let admin = true;
+
 // <!-- ENDPOINTS PARA LA RUTA DE PRODUCTOS -->
 
 //GET para listar todos los productos o por id.-
@@ -25,16 +27,24 @@ routerProductos.get('/:id?', async(req, res)=>{
 })
 //POST para crear nuevo producto
 routerProductos.post('/', async(req, res)=>{
+    
     let newProducto = req.body
-    newProducto.price = parseInt(newProducto.price)
-    let id = await container.save(newProducto)
-    res.status(200).json({"Estado:":`Se agrego correctamente el nuevo producto con nro de id: ${id}`})
+    console.log(newProducto)
+    if(newProducto && newProducto.nombre && newProducto.descripcion && newProducto.foto && newProducto.codigo && newProducto.precio && newProducto.stock){
+        newProducto.precio = parseInt(newProducto.precio) //ver aca!
+        newProducto.stock = parseInt(newProducto.stock)
+        let id = await container.save(newProducto)
+        res.status(200).json({Success:`Se agrego correctamente el nuevo producto con ID: ${id}`})
+    }else{
+        res.status(200).json({Error:`No se pudo agregar el producto, verifique los datos`})
+    }
+    
 })
 //PUT para editar un producto por id
 routerProductos.put('/:id', async(req, res)=>{
     try{
         let data = await container.editById(req.params.id, req.body)
-        res.status(200).json({status: data})
+        res.status(200).json({Success: data})
     }
     catch(err){
         res.status(400).json({"error": err})
@@ -42,13 +52,17 @@ routerProductos.put('/:id', async(req, res)=>{
 })
 //DELETE por id
 routerProductos.delete('/:id', async (req, res)=>{
-    try{
-        let id = parseInt(req.params.id)
-        await container.deleteById(id)
-        res.status(200).json({status: `Se borro el producto con id: ${id}`})
-    }
-    catch(error){
-        res.json({"error": error})
+    if(admin){
+        try{
+            let id = parseInt(req.params.id)
+            await container.deleteById(id)
+            res.status(200).json({deleted: true})
+        }
+        catch(error){
+            res.json({error: error})
+        }
+    }else{
+        res.json({deleted: false})
     }
 })
